chore: reorganize test folder structure
