+++ __tests__/lib/mdx/exports/input/multipleExports.mdx
