+++ __tests__/lib/mdx/exports/input/singleExport.mdx
