+++ __tests__/lib/mdx/exports/input/weirdExports.mdx
