+++ __tests__/lib/mdx/mdast/images/inline/in.mdx
