+++ __tests__/lib/mdx/mdast/images/inline/out.json
