+++ __tests__/lib/mdx/mdast/null-attributes/out.json
