+++ __tests__/lib/mdx/mdast/variables-with-spaces/out.json
