+++ __tests__/lib/mdx/mdast/variables/out.json
