+++ __tests__/lib/mdxish/hast.test.ts
@@ -1,6 +1,6 @@
 import { h } from 'hastscript';
 
-import { hast } from '../../lib';
+import { hast } from '../../../lib';
 
 describe('hast transformer', () => {
   it('parses components into the tree', () => {
